TextField: render helperText if no error
If one passes a `helperText` to `TextField` it should actually be
rendered if no error is present. If there is an error, the error message
should be rendered instead.

The TypeScript compiler complained that `FieldRenderProps` is a generic
type but was used without type information. So I've added the types of
the corresponding `HTML*Element`s.

While at it I've also added the properties of the corresponding MUI
components to the properties to make the `...rest` properties type-safe.

I've exported the properties together with their components.

Finally I've added `.editorconfig` and `.prettierrc` to make formatting
more consistent.